Move missing_command and missing_root_dir tests to runtime test_runtime_errors as these tests are also runtime errors.
This also removes additional test files.

Changes to be committed:
	deleted:    tests/test_missing_command.py
	deleted:    tests/test_missing_root_dir.py
	modified:   tests/test_runtime_errors.py
